squash! WIP: Add a six-column QWERTY and Swedish layout
Alter top row in MIRYOKU_NAV=VI to how it currently is (but should not be?)

[docs/reference/readme.org#vi-style](https://github.com/manna-harbour/miryoku/blob/master/docs/reference/readme.org#vi-style) states: "Columns are moved left one position to match ~hjkl~ positions, and the inner index column is moved to the pinkie column.  All same hand sub layers are modified in the same way" whereas [miryoku_layer_alternatives.h#L272](https://github.com/manna-harbour/miryoku_qmk/blob/miryoku/users/manna-harbour_miryoku/miryoku_babel/miryoku_layer_alternatives.h#L272) actually is `U_RDO, U_PST, U_CPY, U_CUT, U_UND`, same as the other alternatives.

Just alter the layout illustrations for now, until it can be properly understood and perhaps changed in implementation.

Thanks to [Kamal Al-Wendawi](https://www.linkedin.com/in/kamal-al-wendawi-236509158) for diligently helping me troubleshoot my layout illustrations.

TODO: Regenerate the renderings