Redo KMonad Lite layouts according to comments
In #106 Manna Harbour alerted me to the intended use of Lite layout variations - to enable mod taps and the navigation layer without requiring the thumb keys from ortho keyboards. Rather, Lite layouts are intended to provide the minimum viable additions to standard keyboards, so in this version I have only kept those legends.

Also, rather than bothering with the Miryoku theme (necessary to keep all those layer legend variations legible), these diagrams are additions to the KLE presets for 60% keyboards. I am eyeing the [hysp keymap diagrams](https://github.com/alvaro-prieto/corne) to come up with a more pleasant color theme. Obviously, if one want to merge my contributions it could be necessary to perform also other subsequent changes to align with these. Let me know and perhaps I could do that work.

A minor note - commonly shift layers are illustrated as primary (eg. '1') at the bottom and shifted ('!') above it. In KLE this comes out as shifted in the _primary_ position. Also I guessed a bit in terms of how the Miryoku layouts position punctuation keys versus how they are in standard layouts. If one would want to generate these from Bab
el, it would be something to keep in mind.

This time, the screenshots are, as the originals, carefully made pixel perfect in Gimp to maintain the transparent corners of the images.